Copy all the MUSE files to this repo
This is so that I do not have to keep the external disk plugged in all
the time.  I swapped all the linesum images that are currently loaded
in DS9 into the data/muse/ folder and I reloaded them from there into
DS9.  Then I did a backup of current DS9 configuration